requirements.txt
# Core dependencies
openai>=1.10.0
aiohttp>=3.8.5
python-dotenv>=1.0.0

# Canvas API interaction
fastapi>=0.103.1
uvicorn>=0.23.2

# Terminal interface
prompt-toolkit>=3.0.39

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1 